+++ app/controllers/restaurants_controller.rb
@@ -9,7 +9,7 @@ class RestaurantsController < ApplicationController
       {
         lat: restaurant.latitude,
         lng: restaurant.longitude,
-        infoWindow: render_to_string(partial: "infowindow", locals: { restaurant: restaurant })
+        infoWindow: {content: render_to_string(partial: "/restaurants/map_box", locals: { restaurant: restaurant }) }
       }
     end
   end
