+++ app/javascript/packs/application.js
@@ -1,10 +1,6 @@
 import "bootstrap";
 import { initSweetalert } from '../plugins/init_sweetalert';
-import 'mapbox-gl/dist/mapbox-gl.css'; // <-- you need to uncomment the stylesheet_pack_tag in the layout!
 
-import { initMapbox } from '../plugins/init_mapbox';
-
-initMapbox();
 
 initSweetalert('#sweet-alert-demo', {
   title: "Are you sure?",
