+++ app/javascript/packs/map.js
@@ -1,0 +1,38 @@
+import 'mapbox-gl/dist/mapbox-gl.css'
+import '@mapbox/mapbox-gl-geocoder/dist/mapbox-gl-geocoder.css';
+import mapboxgl from 'mapbox-gl/dist/mapbox-gl.js';
+import MapboxGeocoder from '@mapbox/mapbox-gl-geocoder';
+
+  const mapElement = document.getElementById('map');
+
+  if (mapElement) { // only build a map if there's a div#map to inject into
+    mapboxgl.accessToken = process.env.MAPBOX_API_KEY;
+    const map = new mapboxgl.Map({
+      container: 'map',
+      style: 'mapbox://styles/mapbox/streets-v10'
+    });
+
+    const markers = JSON.parse(mapElement.dataset.markers);
+
+    markers.forEach((marker) => {
+      new mapboxgl.Marker()
+        .setLngLat([marker.lng, marker.lat])
+        .setPopup(new mapboxgl.Popup({ offset: 25 })
+        .setHTML(marker.infoWindow.content))
+        .addTo(map);
+    })
+
+   if (markers.lenght === 0) {
+     map.setZoom(1);
+    } else if (markers.lenght === 1) {
+      map.setZoom(14);
+      map.setCenter([markers[0].lng, marker[0].lat]);
+    } else {
+      const bounds = new mapboxgl.LngLatBounds();
+      markers.forEach((marker) => {
+        bounds.extend([marker.lng, marker.lat]);
+      });
+      map.fitBounds(bounds, { duration: 0, padding: 75 })
+    }
+
+  }
