+++ app/views/restaurants/_map_box.html.erb
@@ -1,0 +1,3 @@
+<img src="<%= restaurant.photo1 %>" alt="">
+<h2> <%= restaurant.name %></h2>
+<p><%= restaurant.address %></p>
